Separate ArticleGenerator to a new file as we can use it as external module
